refractor code

+++ CSAT_BMTT/Controllers/UsersController.cs
@@ -16,12 +16,9 @@ namespace CSAT_BMTT.Controllers
     {
         private readonly CSAT_BMTTContext _context;
 
-        public UsersController(CSAT_BMTTContext context)
-        {
-            _context = context;
-        }
+        public UsersController(CSAT_BMTTContext context) => _context = context;
 
-        [HttpGet("")]
+        [HttpGet]
         public async Task<IActionResult> Index(string? searchString)
         {
             if (_context.User == null)
@@ -29,36 +26,22 @@ namespace CSAT_BMTT.Controllers
                 return Problem("Entity set 'CSAT_BMTTContext.User' is null.");
             }
 
-            var currentID = User.FindFirstValue(ClaimTypes.NameIdentifier);
-            var currentEncryptedUser = await _context.User
-                .FirstOrDefaultAsync(u => u.Id.ToString() == currentID);
-
-            var pinCode = "190103";
-            var pinCodesKey = pinCode + currentEncryptedUser.CitizenIdentificationNumber[..10];
-            var pinCodeIv = string.Concat(Enumerable.Repeat(pinCode, 9)) + currentEncryptedUser.CitizenIdentificationNumber[..10];
-
-            var decryptedPrivateKey = AesHelper.Decrypt(currentEncryptedUser.PrivateKey, pinCodeIv, pinCodesKey);
-            var decryptedIvKey = RsaHelper.Decrypt(currentEncryptedUser.IvKey, decryptedPrivateKey);
-            var decryptedStaticKey = RsaHelper.Decrypt(currentEncryptedUser.StaticKey, decryptedPrivateKey);
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var currentUser = await _context.User.FindAsync(int.Parse(currentUserId));
+            if (currentUser == null) return Unauthorized();
 
             var currentDecryptedUser = new User
             {
-                Id = currentEncryptedUser.Id,
-                UserName = AesHelper.Decrypt(currentEncryptedUser.CitizenIdentificationNumber.ToString(), decryptedIvKey, decryptedStaticKey),
-                CitizenIdentificationNumber = currentEncryptedUser.CitizenIdentificationNumber.ToString(),
-                Adress = AesHelper.Decrypt(currentEncryptedUser.Adress, decryptedIvKey, decryptedStaticKey),
-                ATM = AesHelper.Decrypt(currentEncryptedUser.ATM.ToString(), decryptedIvKey, decryptedStaticKey),
-                Birthday = AesHelper.Decrypt(currentEncryptedUser.Birthday.ToString(), decryptedIvKey, decryptedStaticKey),
-                Email = AesHelper.Decrypt(currentEncryptedUser.Email, decryptedIvKey, decryptedStaticKey),
-                Name = currentEncryptedUser.Name,
-                PhoneNumber = AesHelper.Decrypt(currentEncryptedUser.PhoneNumber.ToString(), decryptedIvKey, decryptedStaticKey),
+                Id = currentUser.Id,
+                CitizenIdentificationNumber = currentUser.CitizenIdentificationNumber,
+                Name = currentUser.Name
             };
 
-            var users = await _context.User.Where(u => u.Id.ToString() != currentID).ToListAsync();
+            var users = await _context.User.Where(u => u.Id.ToString() != currentUserId).ToListAsync();
 
-            if (!String.IsNullOrEmpty(searchString))
+            if (!string.IsNullOrEmpty(searchString))
             {
-                users = users.Where(s => s.Name!.ToUpper().Contains(searchString.ToUpper())).ToList();
+                users = users.Where(u => u.Name.Contains(searchString, StringComparison.OrdinalIgnoreCase)).ToList();
             }
 
             var model = new UsersViewModel
@@ -73,55 +56,22 @@ namespace CSAT_BMTT.Controllers
         [HttpGet("details/{id}")]
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var user = await _context.User.FirstOrDefaultAsync(m => m.Id == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-
+            if (user == null) return NotFound();
             return View(user);
         }
 
         [HttpGet("edit/{id}")]
         public async Task<IActionResult> Edit(int? id, string pinCode)
         {
-            if (id == null || String.IsNullOrEmpty(pinCode))
-            {
-                return NotFound();
-            }
+            if (id == null || String.IsNullOrEmpty(pinCode)) return NotFound();
 
             var encryptedUser = await _context.User.FindAsync(id);
-            if (encryptedUser == null)
-            {
-                return NotFound();
-            }
+            if (encryptedUser == null) return NotFound();
 
             // Thêm check quyền
-
-            //var pinCode = "190103";
-            var pinCodesKey = pinCode + encryptedUser.CitizenIdentificationNumber[..10];
-            var pinCodeIv = string.Concat(Enumerable.Repeat(pinCode, 9)) + encryptedUser.CitizenIdentificationNumber[..10];
-
-            var decryptedPrivateKey = AesHelper.Decrypt(encryptedUser.PrivateKey, pinCodeIv, pinCodesKey);
-            var decryptedIvKey = RsaHelper.Decrypt(encryptedUser.IvKey, decryptedPrivateKey);
-            var decryptedStaticKey = RsaHelper.Decrypt(encryptedUser.StaticKey, decryptedPrivateKey);
-
-            var currentDecryptedUser = new User
-            {
-                UserName = AesHelper.Decrypt(encryptedUser.CitizenIdentificationNumber, decryptedIvKey, decryptedStaticKey),
-                CitizenIdentificationNumber = encryptedUser.CitizenIdentificationNumber,
-                Adress = AesHelper.Decrypt(encryptedUser.Adress, decryptedIvKey, decryptedStaticKey),
-                ATM = AesHelper.Decrypt(encryptedUser.ATM, decryptedIvKey, decryptedStaticKey),
-                Birthday = AesHelper.Decrypt(encryptedUser.Birthday, decryptedIvKey, decryptedStaticKey),
-                Email = AesHelper.Decrypt(encryptedUser.Email, decryptedIvKey, decryptedStaticKey),
-                Name = encryptedUser.Name,
-                PhoneNumber = AesHelper.Decrypt(encryptedUser.PhoneNumber, decryptedIvKey, decryptedStaticKey),
-            };
+            var currentDecryptedUser = DecryptUser(encryptedUser, pinCode);
 
             return View(new UserDto(currentDecryptedUser, ""));
         }
@@ -130,32 +80,15 @@ namespace CSAT_BMTT.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [FromForm] UserDto userModel)
         {
-            if (id != userModel.Id)
-            {
-                return NotFound();
-            }
+            if (id != userModel.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
                 var user = await _context.User.FindAsync(id);
-                if (user == null)
-                {
-                    return NotFound();
-                }
+                if (user == null) return NotFound();
+                
+                EncryptUser(user, userModel);
 
-                var pinCodesKey = userModel.PinCode + userModel.CitizenIdentificationNumber[..10];
-                var pinCodeIv = string.Concat(Enumerable.Repeat(userModel.PinCode, 9)) + userModel.CitizenIdentificationNumber[..10];
-
-                var decryptedPrivateKey = AesHelper.Decrypt(user.PrivateKey, pinCodeIv, pinCodesKey);
-                var decryptedIvKey = RsaHelper.Decrypt(user.IvKey, decryptedPrivateKey);
-                var decryptedStaticKey = RsaHelper.Decrypt(user.StaticKey, decryptedPrivateKey);
-
-                user.Adress = AesHelper.Encrypt(userModel.Adress, decryptedIvKey, decryptedStaticKey);
-                user.ATM = AesHelper.Encrypt(userModel.ATM, decryptedIvKey, decryptedStaticKey);
-                user.Birthday = AesHelper.Encrypt(userModel.Birthday, decryptedIvKey, decryptedStaticKey);
-                user.Email = AesHelper.Encrypt(userModel.Email, decryptedIvKey, decryptedStaticKey);
-                user.Name = userModel.Name;
-                user.PhoneNumber = AesHelper.Encrypt(userModel.PhoneNumber, decryptedIvKey, decryptedStaticKey);
                 try
                 {
                     _context.Update(user);
@@ -180,18 +113,9 @@ namespace CSAT_BMTT.Controllers
         [HttpGet("delete/{id}")]
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var user = await _context.User
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var user = await _context.User.FirstOrDefaultAsync(m => m.Id == id);
+            if (user == null) return NotFound();
             return View(user);
         }
 
@@ -200,11 +124,7 @@ namespace CSAT_BMTT.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var user = await _context.User.FindAsync(id);
-            if (user != null)
-            {
-                _context.User.Remove(user);
-            }
-
+            if (user != null) _context.User.Remove(user);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
@@ -212,6 +132,46 @@ namespace CSAT_BMTT.Controllers
         private bool UserExists(int id)
         {
             return _context.User.Any(e => e.Id == id);
+        }
+
+        private (string IvKey, string StaticKey) DecryptKeys(User user, string pinCode)
+        {
+            var pinCodesKey = pinCode + user.CitizenIdentificationNumber[..10];
+            var pinCodeIv = string.Concat(Enumerable.Repeat(pinCode, 9)) + user.CitizenIdentificationNumber[..10];
+            var decryptedPrivateKey = AesHelper.Decrypt(user.PrivateKey, pinCodeIv, pinCodesKey);
+            return
+            (
+                IvKey: RsaHelper.Decrypt(user.IvKey, decryptedPrivateKey),
+                StaticKey: RsaHelper.Decrypt(user.StaticKey, decryptedPrivateKey)
+            );
+        }
+
+        private User DecryptUser(User encryptedUser, string pinCode)
+        {
+            var decryptedKeys = DecryptKeys(encryptedUser, pinCode);
+            return new User
+            {
+                UserName = AesHelper.Decrypt(encryptedUser.CitizenIdentificationNumber, decryptedKeys.IvKey, decryptedKeys.StaticKey),
+                CitizenIdentificationNumber = encryptedUser.CitizenIdentificationNumber,
+                Adress = AesHelper.Decrypt(encryptedUser.Adress, decryptedKeys.IvKey, decryptedKeys.StaticKey),
+                ATM = AesHelper.Decrypt(encryptedUser.ATM, decryptedKeys.IvKey, decryptedKeys.StaticKey),
+                Birthday = AesHelper.Decrypt(encryptedUser.Birthday, decryptedKeys.IvKey, decryptedKeys.StaticKey),
+                Email = AesHelper.Decrypt(encryptedUser.Email, decryptedKeys.IvKey, decryptedKeys.StaticKey),
+                Name = encryptedUser.Name,
+                PhoneNumber = AesHelper.Decrypt(encryptedUser.PhoneNumber, decryptedKeys.IvKey, decryptedKeys.StaticKey),
+            };
+        }
+
+        private void EncryptUser(User user, UserDto userModel)
+        {
+            var decryptedKeys = DecryptKeys(user, userModel.PinCode);
+
+            user.Adress = AesHelper.Encrypt(userModel.Adress, decryptedKeys.IvKey, decryptedKeys.StaticKey);
+            user.ATM = AesHelper.Encrypt(userModel.ATM, decryptedKeys.IvKey, decryptedKeys.StaticKey);
+            user.Birthday = AesHelper.Encrypt(userModel.Birthday, decryptedKeys.IvKey, decryptedKeys.StaticKey);
+            user.Email = AesHelper.Encrypt(userModel.Email, decryptedKeys.IvKey, decryptedKeys.StaticKey);
+            user.Name = userModel.Name;
+            user.PhoneNumber = AesHelper.Encrypt(userModel.PhoneNumber, decryptedKeys.IvKey, decryptedKeys.StaticKey);
         }
     }
 }
